requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.2
cssselect==1.2.0
defusedxml==0.8.0rc2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
fake-useragent==1.4.0
idna==3.6
importlib-metadata==7.0.1
lxml==5.1.0
Markdown==3.5.1
mysqlclient==2.2.1
oauthlib==3.2.2
parse==1.20.1
pycparser==2.21
pyee==8.2.2
PyJWT==2.8.0
pyppeteer==1.0.2
pyquery==2.0.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.4
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.17.0
